backend/requirements.txt
aiobotocore==2.4.2
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.17.0
altair==4.2.0
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
beanie==1.16.4
black==22.12.0
blinker==1.5
boto3==1.24.59
botocore==1.27.59
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
databases==0.6.1
decorator==5.1.1
Django==4.1.4
dnspython==2.2.1
docstring-parser==0.12
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
fastapi==0.88.0
fastapi-pagination==0.11.1
frozenlist==1.3.3
fsspec==2022.10.0
geomet==0.2.1.post1
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lazy-model==0.0.5
MarkupSafe==2.1.1
mmhash3==3.0.1
mongoengine==0.24.2
motor==3.1.1
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
orjson==3.8.3
orm==0.3.1
ormar==0.12.0
packaging==22.0
pandas==1.5.2
passlib==1.7.4
pathspec==0.10.3
piccolo==0.101.0
Pillow==9.4.0
platformdirs==2.6.0
pony==0.7.16
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.14.0
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.0.9
pypika-tortoise==0.1.6
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
s3fs==2022.10.0
s3transfer==0.6.0
scylla-driver==3.25.10
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlakeyset==1.0.1659142803
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a31
sqlmodel==0.0.8
sqlparse==0.4.3
starlette==0.22.0
streamlit==1.16.0
targ==0.3.7
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tortoise-orm==0.19.2
typesystem==0.3.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
zstandard==0.19.0
psycopg2-binary
